refactor: Change type/interface files structure individually.
