+++ src/components/PortfolioCard.tsx
@@ -1,4 +1,4 @@
-import { PortfolioCase } from "@/lib/types";
+import { PortfolioCase } from "@/types";
 import Link from "next/link";
 import { Card, CardContent, CardHeader } from "./ui/card";
 import Image from "next/image";
