--- src/lib/types.ts
@@ -1,94 +1,0 @@
-// Slides
-export interface SlideResponse {
-  slidesFg: {
-    caption: string;
-    title: string;
-    description: string;
-    buttonText: string;
-    buttonLink: string;
-    image: {
-      node: {
-        sourceUrl: string;
-        altText: string;
-      };
-    };
-    active: boolean;
-    order: number;
-  };
-}
-
-// Products
-export interface Product {
-  id: string;
-  title: string;
-  slug: string;
-  content: string;
-  productsFg: {
-    buttonText: string;
-    neededDescription: string;
-    price: string;
-    subtitle: string;
-    frontIcon: {
-      node: {
-        altText: string;
-        sourceUrl: string;
-      };
-    };
-  };
-  featuredImage: {
-    node: {
-      sourceUrl: string;
-      altText: string;
-    };
-  };
-  productCategories?: {
-    nodes: {
-      name: string;
-      slug: string;
-    }[];
-  };
-}
-
-export interface ProductsResponse {
-  products: {
-    nodes: Product[];
-  };
-}
-
-// Portfolio
-export interface PortfolioCase {
-  id: string;
-  title: string;
-  slug: string;
-  featuredImage: {
-    node: {
-      sourceUrl: string;
-      altText: string;
-    };
-  };
-  portfolioCategories?: {
-    nodes: {
-      name: string;
-      slug: string;
-    }[];
-  };
-  portfolioFg: {
-    relatedProduct: {
-      nodes: {
-        id: string;
-      }[];
-    };
-  };
-}
-
-export interface PageInfo {
-  hasNextPage: boolean;
-  endCursor: string;
-}
-
-export interface PaginatedPortfoliosResponse {
-  portfolios: {
-    nodes: PortfolioCase[];
-    pageInfo: PageInfo;
-  };
-}
