+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./slides";
+export * from "./portfolio";
+export * from "./products";
